Change the serializer/deserializer from XStream to Jackson
The point is that we can now avoid saving the null fields(less information to store) and the fact that XStream was an old tool with vulnerability issues.
The point is that we can now avoid saving the null fields (less information to store) and the fact that XStream was an old tool with vulnerability issues.